refactor(bpane-host): split tile_loop/run.rs into per-phase method files
Decompose the 1360-line run() loop into 5 focused processing phases,
each in its own file with a clear method signature:

- cdp_scroll.rs (496 lines): process_cdp_and_scroll() — CDP hints,
  editable regions, scroll detection, trust resolution
- classify.rs (243 lines): classify_tiles() — two-pass video tile
  classification with motion features and hysteresis
- scroll_emit.rs (466 lines): compute_scroll_and_dirty() — residual
  analysis, thin mode, dirty set computation
- emit.rs (193 lines): emit_tiles() — static/content split, tile
  encoding, frame send
- h264_region.rs (111 lines): update_h264_region() — region commit,
  jitter filter, H.264 toggle

run.rs is now 212 lines — a clean orchestration loop calling 5 phases.
frame_types.rs defines CdpScrollResult, ClassifyResult, ScrollEmitResult,
DetectedScrollFrame structs for inter-phase data flow.

Added 3 tests for result struct construction. All 303 tests pass.
